chore: moving `initialize_observability` around
