+++ crates/wash-runtime/Cargo.toml
@@ -59,12 +59,18 @@ tonic = { workspace = true, features = [
     "codegen",
     "tls-native-roots",
 ] }
+tracing-subscriber = { workspace = true, features = ["env-filter", "ansi", "time", "json"] }
+tracing-opentelemetry = { workspace = true }
 pbjson = { workspace = true, default-features = true }
 pbjson-types = { workspace = true, default-features = true }
 prost = { workspace = true, default-features = true }
 tonic-prost = { workspace = true, default-features = true }
 opentelemetry = { workspace = true, features = ["trace", "metrics", "logs"] }
+opentelemetry-appender-tracing = { workspace = true }
+opentelemetry_sdk = { workspace = true }
 opentelemetry-http = { workspace = true }
+opentelemetry-otlp = { workspace = true, features = ["grpc-tonic", "trace", "logs"] }
+opentelemetry-semantic-conventions = { workspace = true, features = ["semconv_experimental"] }
 wasi-graphics-context-wasmtime = { git = "https://github.com/wasi-gfx/wasi-gfx-runtime.git", rev = "067d92d0eabd264e15486b9cecbb9f7020257b7b", optional = true }
 wasi-webgpu-wasmtime = { git = "https://github.com/wasi-gfx/wasi-gfx-runtime.git", rev = "067d92d0eabd264e15486b9cecbb9f7020257b7b", optional = true }
