+++ crates/wash-runtime/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod engine;
 pub mod host;
+pub mod observability;
 pub mod plugin;
 pub mod types;
 pub mod wit;
