+++ crates/wash-runtime/src/observability.rs
@@ -1,0 +1,125 @@
+use anyhow::Context;
+
+use opentelemetry::{KeyValue, trace::TracerProvider};
+use opentelemetry_appender_tracing::layer::OpenTelemetryTracingBridge;
+use opentelemetry_sdk::Resource;
+use opentelemetry_semantic_conventions::resource;
+use tracing::Level;
+use tracing_subscriber::{
+    EnvFilter, Layer, Registry, filter::Directive, layer::SubscriberExt, util::SubscriberInitExt,
+};
+
+/// Initialize observability, setting up console & OpenTelemetry layers.
+///
+/// Returns a shutdown function that should be called on process exit to flush any remaining spans/logs
+pub fn initialize_observability(
+    log_level: Level,
+    ansi_colors: bool,
+    verbose: bool,
+) -> anyhow::Result<Box<dyn FnOnce()>> {
+    // STDERR logging layer
+    let mut fmt_filter =
+        EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new(log_level.as_str()));
+    if !verbose {
+        // async_nats prints out on connect
+        fmt_filter = fmt_filter
+            .add_directive(directive("async_nats=error")?)
+            // wasm_pkg_client/core are a little verbose so we set them to error level in non-verbose mode
+            .add_directive(directive("wasm_pkg_client=error")?)
+            .add_directive(directive("wasm_pkg_core=error")?);
+    }
+
+    let fmt_layer = tracing_subscriber::fmt::layer()
+        .with_writer(std::io::stderr)
+        .with_level(true)
+        .with_target(verbose)
+        .with_thread_ids(verbose)
+        .with_thread_names(verbose)
+        .with_file(verbose)
+        .with_line_number(verbose)
+        .with_ansi(ansi_colors)
+        .with_filter(fmt_filter);
+
+    let otel_enabled = std::env::vars().any(|(key, _)| key.starts_with("OTEL_"));
+    if !otel_enabled {
+        Registry::default().with(fmt_layer).init();
+
+        // No-op shutdown function
+        let shutdown_fn = || {};
+        return Ok(Box::new(shutdown_fn));
+    }
+
+    let resource = Resource::builder()
+        .with_attribute(KeyValue::new(
+            resource::SERVICE_NAME.to_string(),
+            env!("CARGO_PKG_NAME"),
+        ))
+        .with_attribute(KeyValue::new(
+            resource::SERVICE_INSTANCE_ID.to_string(),
+            uuid::Uuid::new_v4().to_string(),
+        ))
+        .with_attribute(KeyValue::new(
+            resource::SERVICE_VERSION.to_string(),
+            env!("CARGO_PKG_VERSION"),
+        ))
+        .build();
+
+    // OTel logging layer
+    let log_exporter = opentelemetry_otlp::LogExporter::builder()
+        .with_tonic()
+        .build()?;
+    let log_provider = opentelemetry_sdk::logs::LoggerProviderBuilder::default()
+        .with_batch_exporter(log_exporter)
+        .with_resource(resource.clone())
+        .build();
+    let filter_otel_logs = EnvFilter::new(log_level.as_str());
+
+    let otel_logs_layer =
+        OpenTelemetryTracingBridge::new(&log_provider).with_filter(filter_otel_logs);
+
+    // OTel tracing layer
+    let tracer_exporter = opentelemetry_otlp::SpanExporter::builder()
+        .with_tonic()
+        .build()?;
+    let tracer_provider = opentelemetry_sdk::trace::TracerProviderBuilder::default()
+        .with_batch_exporter(tracer_exporter)
+        .with_resource(resource.clone())
+        .build();
+
+    let filter_otel_traces = EnvFilter::new(log_level.as_str());
+
+    let otel_tracer_layer = tracing_opentelemetry::layer()
+        .with_tracer(tracer_provider.tracer("runtime"))
+        .with_error_records_to_exceptions(true)
+        .with_error_fields_to_exceptions(true)
+        .with_error_events_to_status(true)
+        .with_error_events_to_exceptions(true)
+        .with_location(true)
+        .with_filter(filter_otel_traces);
+
+    Registry::default()
+        .with(fmt_layer)
+        .with(otel_logs_layer)
+        .with(otel_tracer_layer)
+        .init();
+
+    // Return a shutdown function to flush providers on exit
+    let shutdown_fn = move || {
+        if let Err(e) = tracer_provider.shutdown() {
+            eprintln!("failed to shutdown tracer provider: {e}");
+        }
+        if let Err(e) = log_provider.shutdown() {
+            eprintln!("failed to shutdown log provider: {e}");
+        }
+    };
+
+    Ok(Box::new(shutdown_fn))
+}
+
+/// Helper function to reduce duplication and code size for parsing directives
+fn directive(directive: impl AsRef<str>) -> anyhow::Result<Directive> {
+    directive
+        .as_ref()
+        .parse()
+        .with_context(|| format!("failed to parse filter: {}", directive.as_ref()))
+}
