+++ crates/wash/Cargo.toml
@@ -39,11 +39,6 @@ etcetera = { workspace = true }
 figment = { workspace = true, features = ["json", "env", "yaml", "toml"] }
 flate2 = { workspace = true, features = ["rust_backend"] }
 humantime = { workspace = true }
-opentelemetry = { workspace = true }
-opentelemetry_sdk = { workspace = true }
-opentelemetry-semantic-conventions = { workspace = true, features = ["semconv_experimental"] }
-opentelemetry-otlp = { workspace = true, features = ["grpc-tonic", "trace", "logs"] }
-opentelemetry-appender-tracing = { workspace = true }
 reqwest = { workspace = true, features = ["json", "rustls-tls"] }
 rustls = { workspace = true }
 semver = { workspace = true }
@@ -55,7 +50,6 @@ tempfile = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tracing = { workspace = true, features = ["attributes"] }
 tracing-subscriber = { workspace = true, features = ["env-filter", "ansi", "time", "json"] }
-tracing-opentelemetry = { workspace = true }
 url = { workspace = true }
 uuid = { workspace = true }
 wasm-metadata = { workspace = true }
